+++ Godeps/_workspace/src/github.com/emicklei/go-restful/swagger/model_builder.go
@@ -43,6 +43,9 @@ func (b modelBuilder) addModel(st reflect.Type, nameOverride string) {
 	for i := 0; i < st.NumField(); i++ {
 		field := st.Field(i)
 		jsonName, prop := b.buildProperty(field, &sm, modelName)
+		if descTag := field.Tag.Get("description"); descTag != "" {
+			prop.Description = descTag
+		}
 		// add if not ommitted
 		if len(jsonName) != 0 {
 			// update Required
@@ -80,7 +83,6 @@ func (b modelBuilder) buildProperty(field reflect.StructField, model *Model, mod
 	if jsonTag := field.Tag.Get("json"); jsonTag != "" {
 		s := strings.Split(jsonTag, ",")
 		if len(s) > 1 && s[1] == "string" {
-			prop.Description = "(" + fieldType.String() + " as string)"
 			fieldType = reflect.TypeOf("")
 		}
 	}
@@ -139,7 +141,17 @@ func (b modelBuilder) buildStructTypeProperty(field reflect.StructField, jsonNam
 		subModel := sub.Models[subKey]
 		for k, v := range subModel.Properties {
 			model.Properties[k] = v
-			model.Required = append(model.Required, k)
+			// if subModel says this property is required then include it
+			required := false
+			for _, each := range subModel.Required {
+				if k == each {
+					required = true
+					break
+				}
+			}
+			if required {
+				model.Required = append(model.Required, k)
+			}
 		}
 		// empty name signals skip property
 		return "", prop